DELIA-48856: use open source repo for BleRcuDaemon
Reason for change:  changes to get it to compile successfully,
	also pulling in missing changes from other repo:

344111c RDKTV-3810: BleRcuDaemon need to generate minidump
49b1662 RDKTV-4253: Remote pairing failed after a factory reset of TV
68f7f1c RDK-31248,RDKTV-1736: Implement BLE Remote Control Service on GATT client

Test Procedure:     compiles and runs without regressions
Risks:              Low

+++ daemon/source/blercu/bleservices/blercuaudioservice.h
@@ -62,8 +62,8 @@ public:
 public:
 	virtual bool isStreaming() const = 0;
 
-	virtual qint32 gainLevel() const = 0;
-	virtual void setGainLevel(qint32 level) = 0;
+	virtual quint8 gainLevel() const = 0;
+	virtual void setGainLevel(quint8 level) = 0;
 
 	virtual Future<FileDescriptor> startStreaming(Encoding encoding) = 0;
 	virtual Future<> startStreamingTo(Encoding encoding, int pipeWriteFd) = 0;
@@ -86,7 +86,7 @@ public:
 
 signals:
 	void streamingChanged(bool streamming);
-	void gainLevelChanged(qint32 level);
+	void gainLevelChanged(quint8 level);
 };
 
 Q_DECLARE_METATYPE(BleRcuAudioService::StatusInfo)
